+++ src/ui/components/Group.ts
@@ -9,7 +9,10 @@ export interface GroupProps {
 }
 
 function escapeAttr(s: string): string {
-    return s.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/</g, '&lt;');
+    return s
+        .replace(/&/g, '&amp;')
+        .replace(/"/g, '&quot;')
+        .replace(/</g, '&lt;');
 }
 
 export function Group({